+++ CHANGELOG.md
@@ -4,6 +4,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.7.0 - in development
+
+### Changed
+
+- Added an `add_trial_division_test` argument to `fips::is_prime()`. ([#99])
+- `fips::is_prime()` now takes a `FipsOptions` structure for additional options. ([#100])
+
+
+[#99]: https://github.com/entropyxyz/crypto-primes/pull/99
+[#100]: https://github.com/entropyxyz/crypto-primes/pull/100
+
+
 ## [0.7.0-pre.4] - 2025-11-06
 
 ### Changed